Address review: move payloadAvailability to own file, generalize skip log, tighten deadline snapshot
