+++ beacon-chain/rpc/core/validator.go
@@ -956,7 +956,6 @@ func (s *Service) PayloadAttestationData(
 	}
 	cfg := params.BeaconConfig()
 	deadline := slotStart.Add(cfg.SlotComponentDuration(cfg.PayloadAttestationDueBPS))
-	beforeDeadline := prysmTime.Now().Before(deadline)
 
 	if cached := s.payloadAttestationData.Load(); cached != nil && cached.Slot == slot {
 		return cached, nil
@@ -973,7 +972,7 @@ func (s *Service) PayloadAttestationData(
 		// Before the deadline only the final result is safe to return: the payload
 		// arrived timely and the data is available. Otherwise both flags may still
 		// flip, so wait for the deadline.
-		if beforeDeadline && !(data.PayloadPresent && data.BlobDataAvailable) {
+		if prysmTime.Now().Before(deadline) && !(data.PayloadPresent && data.BlobDataAvailable) {
 			return &RpcError{Reason: Unavailable, Err: fmt.Errorf("payload attestation data not yet final for slot %d", slot)}, nil
 		}
 		s.payloadAttestationData.Store(data)
